Add the powershell script, another exemple about '.rpa', and a summary part about workflow.
## Description
The powershell equivalent of the bash script is now added and usable and the exemple as been update.
Another exemple as been add to get the '.rpa' key, since opening a whole '.rpa' file to simply get it is not the most efficiency.
Summaries as been add in the worklow part to show a better overview and the last part of the workflow section as been separate with a 'Also' header.
In addition, some bugs (such as some forgetings) have been fixed.