A bit change the diff command ouput and improve the name-id regex.
## Description
Now, change indications are prefixed by `!` instead of `-` to be more coherent with official diff syntax, and in it's body the 'For' and 'From' respectively represent by the prefix `<` and `>`.

I was forget that the name-ids of 'dialogs' translations can be a string, now it's fix.

Also add some support for comments that can be put at the end of some lines.

A other change is that now '""' translations for '""' does not be count as untranslated anymore.

An another fix about line counting to be a bit less ambiguous.