The name in the dialog identifier is now handled separately with the '--proxy 1' option of populate.
## Description
Now, the 1st proxy level will only compare the part of the identifier _*(with the string itself of course)*_ for the 'populate' command.

This also occasion the catch and fix of an error in the previously lvl 1 proxy.
Also improves extractions of translations (of the module itself).
Fix an error from the previous commit.