Improved readability of choices displayed by the '--ask' option.
## Details
Now, the choices displayed by the '--ask' option are put at the line and indented.
La partie identifiant de nom du dialogue est également affiché.

Also, now, each output file in the 'populate' command is recorded at the end of a population process, allowing to stop without having to start all again from zero after.
Also, if a string is not found in a file, it will be searched for in the other files in the currently processed group.

Fixed a small problem with the 'check' command.