Update the language detection. Also add 2 options.
## Details
I was forget to update the laguage detection for the module help.

Also, move the previously added functionnality to search in other files before trying a proxy to the new '--bunch' option (because it can make the process unnecessarily longer).

Also update the help and the translation of this one (only in french, sorry).