Implementated the --bulk option and add the --where option for the check command and the --what option for the diff command.
## Details
The --bulk option of the populate command is now implemented and can be use to populate few files from many (or many from few).

The --where option can now be use with the check command to see where the empty translations are.
The --what option can now be use with the diff command to see what is add or is no longer present in the translation files.

Also added new regexp identifiers for the post-args and one-line python statement, so they can had more use.
Also update the help and the translations.