Add options about NID and proxy.
## Details
Add the --from-nID, --no-proxy-warn and --proxy-y options.
Add the -- option so we can now indicate where to stop parsing options.
Add the -: option so we can now concate short options (like `-:vd`).
Update scripts to take count of these last 2 adds.

Some other fix *(such as about args of dIDs args that could had be lose in certain specific case on previous versions.)* and improvements.

Add some more info about the --ask option.
Also update the help and the translations (and fix an error with the help for the diff command).